+++ CHANGELOG.md
@@ -24,6 +24,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - metrics: don't create error-report on closing connection nodes
 - LVM: Correctly calculate extents with more than one stripes
+- RscDfn and SnapDfn: Properly restore layerStack (!= layerData) from backup and snapshot restore
 
 ## [1.30.2] - 2024-12-18